# Optimise zlib reader and consolidate sync.Pools