Expands on the optimisations made on https://github.com/go-git/go-git/pull/602 and ensures that zlib reader does not need to recreate a deflate dictionary at every use.

The use of sync pools was consolidated into a new `sync` utils package, for improved code structure.
```
name       old time/op    new time/op    delta
Parser-16    7.51ms ± 3%    7.71ms ± 6%     ~     (p=0.222 n=5+5)

name       old alloc/op   new alloc/op   delta
Parser-16    4.65MB ± 3%    1.90MB ± 3%  -59.06%  (p=0.008 n=5+5)

name       old allocs/op  new allocs/op  delta
Parser-16     3.48k ± 0%     3.32k ± 0%   -4.57%  (p=0.016 n=5+4)
```

:warning: Requires rebase after https://github.com/go-git/go-git/pull/602 is merged.